+++ protobuf/src/well_known_types/field_mask.rs
@@ -142,11 +142,9 @@ impl ::protobuf::Message for FieldMask {
     }
 
     fn descriptor(&self) -> &'static ::protobuf::reflect::MessageDescriptor {
-        ::protobuf::MessageStatic::descriptor_static(None::<Self>)
+        ::protobuf::Message::descriptor_static(None::<Self>)
     }
-}
 
-impl ::protobuf::MessageStatic for FieldMask {
     fn new() -> FieldMask {
         FieldMask::new()
     }